refactor: change messages to websocket domain
